Add 2 new properties for fuzzy positions
It is currently not straight forward to model possible positions for OneOfPosition, see https://github.com/OBF/FALDO/issues/37. The new property possiblePositions makes it simpler.
The second new property for FuzzyPosition, probability, cover cases as those of PTMs in IntAct where a probability is assigned to a fuzzy position.